requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
factory-boy==3.2.1
Faker==15.3.4
fastapi==0.82.0
fastapi-health==0.4.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
motor==3.1.1
odmantic==0.9.1
outcome==1.2.0
packaging==22.0
passlib==1.7.4
pipdeptree==2.3.3
pluggy==1.0.0
pyasn1==0.4.8
pydantic==1.10.2
pymongo==4.3.3
pytest==7.2.0
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.19.1
trio==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4